+++ src/openid/Service.php
@@ -180,19 +180,4 @@ abstract class Service extends ServiceBase implements IAuthService
 			return Yii::$app->getRequest()->getHostInfo();
 		}
 	}
-
-    /**
-     * Returns the public resource.
-     *
-     * @param string $url url to request.
-     * @param array $options HTTP request options. Keys: query, data, headers.
-     * @return mixed the response.
-     * @throws ErrorException
-     */
-    public function makeRequest($url, $options = [])
-    {
-        return $this->request($url, $options, function ($url, $method, $headers, $data) {
-            return $this->getHttpClient()->retrieveResponse($url, $data, $headers, $method);
-        });
-    }
 }