+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Yii2 EAuth Change Log
 =====================
 
+### dev (13.01.2016)
+* Move response parsing from oauth to base service (fix #71)
+
 ### 2.4.0 (03.01.2016)
 * Fixed error param names for Facebook (#63)
 * Use the latest Graph Api v2.5 for Facebook (#65)